Vendor TweaselHar type
While I'm not the biggest fan of duplicating the code, we were also
(needlessly) inheriting the Node 18 restriction from CA. Also, CA is
pretty huge and I never liked depending on that just for this one type.

Since the type is expected to be pretty stable, I don't think vendoring
it is too much of an issue.