Initial implementation (#1)
* Import har2pdf code

* har2pdf: Shift all headings down by one

* har2pdf: Make heading referencable

* First somewhat working prototype

* Description

* Only print requests referenced in the report

* Reference tracker-wiki in report

* Escaping

* Outsource style to external file

* Controller notices

* Complaints

* delete unused translations file

* Fixes #5: Render App Privacy Report/TrackerControl export

* Refactor GenerateOptions

* Mention app version in second analysis in complaint

* Complaint: Only include requests from user's network activity report

* Rename to ReportHAR

* Factor out traffic preparation

* Factor out Nunjucks code

* Factor out Typst compilation

* har2pdf: fix typo in encoding detection

* extract meta from tweasel HARs

* Allow passing harMd5

* Write docstrings

* write readme

* Fixes #4: Include software versions in report

* Add link to HAR tutorial

* Export prepareTraffic()

We need to do the same thing in platform.

* Vendor TweaselHar type

While I'm not the biggest fan of duplicating the code, we were also
(needlessly) inheriting the Node 18 restriction from CA. Also, CA is
pretty huge and I never liked depending on that just for this one type.

Since the type is expected to be pretty stable, I don't think vendoring
it is too much of an issue.

* Upgrade Parcel

Due to: https://github.com/parcel-bundler/parcel/pull/9419

* Switch to our custom nunjucks fork

* Inline TrackHAR translations in bundle

I was getting the following error when trying to use ReportHAR as a
library:

Module "trackhar/i18n/en.json" needs an import assertion of type "json"

But adding the assertion didn't help—Parcel strips those.

* Export translations

* Support TrackerControl individual app exports

* Switch to released TrackHAR

* Template for informal complaints

* Actually convince Parcel to inline the translations and templates

Quite finicky…

* Remove yalc and update TrackHAR

* Add method section to report

---------

Co-authored-by: Lorenz Sieben <me@lorenz-sieben.com>